+++ src/cert_provider/main.cc
@@ -251,7 +251,7 @@ bool generate_and_sign(const std::string& cacert_path, const std::string& capkey
     return false;
   }
   auto privkey_len = BIO_get_mem_data(privkey_file.get(), &privkey_buf);  // NOLINT
-  *pkey = std::string(privkey_buf, privkey_len);
+  *pkey = std::string(privkey_buf, static_cast<size_t>(privkey_len));
 
   // serialize certificate
   char* cert_buf;
@@ -266,7 +266,7 @@ bool generate_and_sign(const std::string& cacert_path, const std::string& capkey
     return false;
   }
   auto cert_len = BIO_get_mem_data(cert_file.get(), &cert_buf);  // NOLINT
-  *cert = std::string(cert_buf, cert_len);
+  *cert = std::string(cert_buf, static_cast<size_t>(cert_len));
 
   return true;
 }
